feat: add binary conversion and number manipulation programs
Add three C++ programs:
1. Binary to decimal converter
2. Number reverser with 32-bit integer range check
3. Power of two checker with bitwise operation

+++ 07_problems/prob1.cpp
@@ -2,11 +2,13 @@
 #include <math.h>
 using namespace std;
 
+//Binary to Decimal
+
 int main(){
     int n;
     cout << "Enter the Number : ";
     cin >> n;
-    int ans=0;
+    float ans=0;
     int i=0;
     while(n != 0){
         int bit = n & 1;
